:peach: use publication review entity and add partial refactor of review cataloger
