+++ src/collector/entities/publication_review.py
@@ -1,8 +1,12 @@
 from dataclasses import dataclass
+from dataclasses_json import dataclass_json
 from src.collector.entities.artist import Artist
+from src.collector.entities.release import Release
+from src.collector.entities.review import Review
 from src.common.crypto import calculate_hash
 
 
+@dataclass_json
 @dataclass
 class PublicationReview:
     artist: str
@@ -12,14 +16,13 @@ class PublicationReview:
     date: str
     link: str
 
-    def artist(self) -> Artist:
-        # id = calculate_hash(self.artist)
-        # artist = Artist(id=id, name=self.artist, releases=[])
-        # return artist
-        pass
-
-    def review(self):
-        pass
-
-    def release(self):
-        pass
+    # def artist(self) -> Artist:
+    #     id = calculate_hash(self.artist)
+    #     artist = Artist(id=id, name=self.artist, releases=[])
+    #     return artist
+    #
+    # def review(self) -> Review:
+    #     raise NotImplemented
+    #
+    # def release(self) -> Release:
+    #     raise NotImplemented
